requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
psycopg2-binary==2.9.10
pika==1.3.1
python-dotenv==1.0.0
